+++ data/config/src/main/java/com/hilingual/data/config/repository/ConfigRepository.kt
@@ -19,6 +19,4 @@ import com.hilingual.data.config.model.AppVersionInfo
 
 interface ConfigRepository {
     suspend fun getAppVersionInfo(): Result<AppVersionInfo>
-    suspend fun getIsSplashOnboardingCompleted(): Result<Boolean>
-    suspend fun completeSplashOnboarding(): Result<Unit>
 }
