+++ presentation/auth/build.gradle.kts
@@ -26,5 +26,5 @@ android {
 dependencies {
     implementation(projects.data.auth)
     implementation(projects.data.user)
-    implementation(projects.data.config)
+    implementation(projects.data.onboarding)
 }
