+++ presentation/splash/build.gradle.kts
@@ -27,4 +27,5 @@ dependencies {
     implementation(projects.data.auth)
     implementation(projects.data.user)
     implementation(projects.data.config)
+    implementation(projects.data.onboarding)
 }
